+++ src/games/blam.ts
@@ -4,6 +4,7 @@ import { RectGrid } from "../common";
 import { APRenderRep } from "@abstractplay/renderer/src/schema";
 import { Directions } from "../common";
 import { APMoveResult } from "../schemas/moveresults";
+import { reviver } from "../common/serialization";
 // tslint:disable-next-line: no-var-requires
 const clone = require("rfdc/default");
 
@@ -68,14 +69,14 @@ export class BlamGame extends GameBase {
     public lastmove?: string;
     public gameover: boolean = false;
     public winner: playerid[] = [];
-    public variants?: string[];
+    public variants: string[] = [];
     public scores!: number[];
     public caps!: number[];
     public stashes!: Map<playerid, number[]>;
-    public stack: Array<IMoveState>;
+    public stack!: Array<IMoveState>;
     public results: Array<APMoveResult> = []
 
-    constructor(state: number | IBlamState, variants?: string[]) {
+    constructor(state: number | IBlamState | string, variants?: string[]) {
         super();
         if (typeof state === "number") {
             this.numplayers = state;
@@ -98,6 +99,9 @@ export class BlamGame extends GameBase {
             }
             this.stack = [fresh];
         } else {
+            if (typeof state === "string") {
+                state = JSON.parse(state, reviver) as IBlamState;
+            }
             if (state.game !== BlamGame.gameinfo.uid) {
                 throw new Error(`The Blam! game code cannot process a game of '${state.game}'.`);
             }